+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+### 改变
+
+- build.rs及Makefile 支持条件编译。
+- 补上一个空的 read_stream。
+
 ## [0.1.1] - 2019-3-7
 
 ### 增加